Various changes to enable cross-compilation (without `-fexternal-interpreter`)
As it stands, we also need changes to `entropy`, but I'll manage that separately. Hopefully by reviving https://github.com/TomMD/entropy/pull/20.

Note that the custom setup is only used for `cabal-doctest`, so we really can just safely remove it. This is essentially the same issue as https://github.com/cdepillabout/pretty-simple/issues/82.

Seeing as all the TH is just for `makeLenses`, in an ideal world we really ought to just be able to run on the host: see https://github.com/ghc-proposals/ghc-proposals/pull/243.

The changes to the two `.hs` files are from dumping the TH splices (and manually (well, with _some_ help from HLS) adding a load of qualified imports). The splices can be found in `dist-newstyle` after adding this to `cabal.project`:

```
package wreq
    ghc-options: -ddump-splices -dth-dec-file
```